[#69] Introduce SubscriberMgmt trait to make Sample independent of SubscriberImpl specific generic arguments
