+++ iceoryx2/tests/service_publish_subscribe_tests.rs
@@ -14,7 +14,7 @@
 mod service_publish_subscribe {
     use iceoryx2::config::Config;
     use iceoryx2::port::publisher::{PublisherCreateError, PublisherLoanError};
-    use iceoryx2::port::subscriber_impl::SubscriberCreateError;
+    use iceoryx2::port::subscriber::SubscriberCreateError;
     use iceoryx2::prelude::*;
     use iceoryx2::service::builder::publish_subscribe::PublishSubscribeCreateError;
     use iceoryx2::service::builder::publish_subscribe::PublishSubscribeOpenError;
